HBSD: fix build on i386
MFC-to: 10-STABLE
(cherry picked from commit cc41c6dd9402e4df0ee53cbec36d1ea712da7c47)
(cherry picked from commit d350e83fc4a30970192e7a31d8b26c78e1c87549)

+++ sys/hardenedbsd/hbsd_pax_common.c
@@ -345,7 +345,7 @@ pax_sysinit(void)
 {
 
 	printf("HBSD: initialize and check HardenedBSD features (version %"PRIu64").\n",
-	    __HardenedBSD_version);
+	    (uint64_t)__HardenedBSD_version);
 }
 SYSINIT(pax, SI_SUB_PAX, SI_ORDER_FIRST, pax_sysinit, NULL);
